Worker/requirements.txt
pandas~=2.2.2
folium~=0.16.0
geopy~=2.4.1
SQLAlchemy
#SQLAlchemy~=2.0.30
#SQLAlchemy<1.4.0
requests~=2.31.0
xlrd~=2.0.1
xlwt~=1.3.0
python-dotenv~=1.0.1
#psycopg2-binary
psycopg2~=2.9.9
sympy~=1.13.1